remove deprecated members and extend html methods (#488)
