+++ apps/dart_quotes/lib/pages/quote_page.dart
@@ -24,7 +24,7 @@ class QuotePage extends AsyncStatelessComponent {
       title: quote.quote,
       meta: {"description": '"${quote.quote}" - ${quote.author}'},
       children: [
-        script(src: "https://cdn.jsdelivr.net/npm/js-confetti@latest/dist/js-confetti.browser.js", defer: true, []),
+        script(src: "https://cdn.jsdelivr.net/npm/js-confetti@latest/dist/js-confetti.browser.js", defer: true),
       ],
     );
