+++ apps/dart_quotes_server/lib/web/pages/quote_page.dart
@@ -25,7 +25,7 @@ class QuotePage extends AsyncStatelessComponent {
       title: quote.quote,
       meta: {"description": '"${quote.quote}" - ${quote.author}'},
       children: [
-        script(src: "https://cdn.jsdelivr.net/npm/js-confetti@latest/dist/js-confetti.browser.js", defer: true, []),
+        script(src: "https://cdn.jsdelivr.net/npm/js-confetti@latest/dist/js-confetti.browser.js", defer: true),
       ],
     );
