+++ examples/flutter_embedding/lib/main.dart
@@ -26,7 +26,7 @@ void main() {
       link(rel: 'manifest', href: 'manifest.json'),
       link(rel: 'stylesheet', href: 'css/style.css', type: 'text/css'),
 
-      script(src: "flutter_bootstrap.js", async: true, []),
+      script(src: "flutter_bootstrap.js", async: true),
     ],
     body: App(),
   ));
