+++ examples/flutter_multi_view/lib/main.dart
@@ -12,7 +12,7 @@ void main() {
     title: 'flutter_multi_view',
     head: [
       link(rel: 'manifest', href: 'manifest.json'),
-      script(src: "flutter_bootstrap.js", async: true, []),
+      script(src: "flutter_bootstrap.js", async: true),
     ],
     styles: [
       css.import('https://fonts.googleapis.com/css?family=Roboto'),
