--- packages/jaspr/lib/src/foundation/styles/groups/grid_item.dart
@@ -1,10 +1,0 @@
-part of '../styles.dart';
-
-class _GridItemStyles extends Styles {
-  const _GridItemStyles({this.placement}) : super._();
-
-  final GridPlacement? placement;
-
-  @override
-  Map<String, String> get properties => placement?.styles ?? {};
-}
