+++ packages/jaspr/test/browser/node_key/node_key_test.dart
@@ -10,7 +10,7 @@ void main() {
     testBrowser('should access node from GlobalNodeKey', (tester) async {
       final key = GlobalNodeKey<HTMLInputElement>();
 
-      tester.pumpComponent(input(key: key, type: InputType.checkbox, []));
+      tester.pumpComponent(input(key: key, type: InputType.checkbox));
 
       expect(key.currentNode, isNotNull);
