+++ packages/jaspr/test/framework/basic/basic_app.dart
@@ -38,10 +38,9 @@ class Button extends StatelessComponent {
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield DomComponent(
-      tag: 'button',
-      child: Text(label),
+    yield button(
       events: {'click': (e) => onPressed()},
+      [text(label)],
     );
   }
 }
