+++ packages/jaspr/tool/data/html.json
@@ -510,6 +510,12 @@
           "type": "string"
         }
       }
+    },
+    "figure": {
+      "doc": "The &lt;figure&gt; HTML element represents self-contained content, potentially with an optional caption, which is specified using the &lt;figcaption&gt; element. The figure, its caption, and its contents are referenced as a single unit."
+    },
+    "figcaption": {
+      "doc": "The &lt;figcaption&gt; HTML element represents a caption or legend describing the rest of the contents of its parent &lt;figure&gt; element, providing the &lt;figure&gt;> an accessible description."
     }
   },
   "forms": {
@@ -628,6 +634,7 @@
     },
     "input": {
       "doc": "The &lt;input&gt; HTML element is used to create interactive controls for web-based forms in order to accept data from the user; a wide variety of types of input data and control widgets are available, depending on the device and user agent. The &lt;input&gt; element is one of the most powerful and complex in all of HTML due to the sheer number of combinations of input types and attributes.",
+      "self_closing": true,
       "attributes": {
         "type": {
           "doc": "Defines how an &lt;input&gt; works. If this attribute is not specified, the default type adopted is text.",
@@ -1041,6 +1048,7 @@
     },
     "script": {
       "doc": "The &lt;script&gt; HTML element is used to embed executable code or data; this is typically used to embed or refer to JavaScript code. The &lt;script&gt; element can also be used with other languages, such as WebGL's GLSL shader programming language and JSON.",
+      "self_closing": true,
       "attributes": {
         "async": {
           "doc": "For classic scripts, if the async attribute is present, then the classic script will be fetched in parallel to parsing and evaluated as soon as it is available.\n\nFor module scripts, if the async attribute is present then the scripts and all their dependencies will be executed in the defer queue, therefore they will get fetched in parallel to parsing and evaluated as soon as they are available.\n\nThis attribute allows the elimination of parser-blocking JavaScript where the browser would have to load and evaluate scripts before continuing to parse. defer has a similar effect in this case.",
@@ -1052,8 +1060,11 @@
         },
         "src": {
           "doc": "This attribute specifies the URI of an external script; this can be used as an alternative to embedding a script directly within a document.",
-          "required": true,
           "type": "string"
+        },
+        "content": {
+          "doc": "The content of the script element, if it is not an external script.",
+          "type": "content"
         }
       }
     }
@@ -1277,6 +1288,7 @@
     },
     "col": {
       "doc": "The &lt;col&gt; HTML element defines one or more columns in a column group represented by its parent &lt;colgroup&gt; element. The &lt;col&gt; element is only valid as a child of a &lt;colgroup&gt; element that has no span attribute defined.",
+      "self_closing": true,
       "attributes": {
         "span": {
           "doc": "Specifies the number of consecutive columns the &lt;col&gt; element spans. The value must be a positive integer greater than zero. If not present, its default value is 1.",
