+++ packages/jaspr_builder/test/styles/sources/styles_class.dart
@@ -7,13 +7,13 @@ const stylesClassSources = {
     class Component {
       @css
       static final styles = [
-        css('.main').box(width: 100.px),
+        css('.main').styles(width: 100.px),
       ];
       
       @css
       static List<StyleRule> get styles2 => [
         css('.main', [
-          css('&').box(width: 100.px),
+          css('&').styles(width: 100.px),
         ]),
       ];
     }
