+++ packages/jaspr_builder/test/styles/sources/styles_global.dart
@@ -6,13 +6,13 @@ const stylesGlobalSources = {
         
     @css
     final styles = [
-      css('.main').box(width: 100.px),
+      css('.main').styles(width: 100.px),
     ];
     
     @css
     List<StyleRule> get styles2 => [
       css('.main', [
-        css('&').box(width: 100.px),
+        css('&').styles(width: 100.px),
       ]),
     ];
   ''',
