+++ packages/jaspr_cli/lib/src/bundles/scaffold/scaffold_bundle.dart
@@ -26,7 +26,7 @@ final scaffoldBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "lib/base#main.dart",
       "data":
-          "Ly8gVGhlIGVudHJ5cG9pbnQgZm9yIHRoZSAqKnNlcnZlcioqIGVudmlyb25tZW50LgovLwovLyBUaGUgW21haW5dIG1ldGhvZCB3aWxsIG9ubHkgYmUgZXhlY3V0ZWQgb24gdGhlIHNlcnZlciBkdXJpbmcgcHJlLXJlbmRlcmluZy4KLy8gVG8gcnVuIGNvZGUgb24gdGhlIGNsaWVudCwge3teaHlkcmF0aW9ufX1zZWUgW3dlYi9tYWluLmRhcnRde3svaHlkcmF0aW9ufX17eyNoeWRyYXRpb259fXVzZSB0aGUgQGNsaWVudCBhbm5vdGF0aW9ue3svaHlkcmF0aW9ufX0uCgovLyBTZXJ2ZXItc3BlY2lmaWMgamFzcHIgaW1wb3J0LgppbXBvcnQgJ3BhY2thZ2U6amFzcHIvc2VydmVyLmRhcnQnOwoKLy8gSW1wb3J0cyB0aGUgW0FwcF0gY29tcG9uZW50LgppbXBvcnQgJ2FwcC5kYXJ0JzsKCi8vIFRoaXMgZmlsZSBpcyBnZW5lcmF0ZWQgYXV0b21hdGljYWxseSBieSBKYXNwciwgZG8gbm90IHJlbW92ZSBvciBlZGl0LgppbXBvcnQgJ2phc3ByX29wdGlvbnMuZGFydCc7Cgp2b2lkIG1haW4oKSB7CiAgLy8gSW5pdGlhbGl6ZXMgdGhlIHNlcnZlciBlbnZpcm9ubWVudCB3aXRoIHRoZSBnZW5lcmF0ZWQgZGVmYXVsdCBvcHRpb25zLgogIEphc3ByLmluaXRpYWxpemVBcHAoCiAgICBvcHRpb25zOiBkZWZhdWx0SmFzcHJPcHRpb25zLAogICk7CgogIC8vIFN0YXJ0cyB0aGUgYXBwLgogIC8vCiAgLy8gW0RvY3VtZW50XSByZW5kZXJzIHRoZSByb290IGRvY3VtZW50IHN0cnVjdHVyZSAoPGh0bWw+LCA8aGVhZD4gYW5kIDxib2R5PikKICAvLyB3aXRoIHRoZSBwcm92aWRlZCBwYXJhbWV0ZXJzIGFuZCBjb21wb25lbnRzLgogIHJ1bkFwcChEb2N1bWVudCgKICAgIHRpdGxlOiAne3tuYW1lfX0nLAogICAgc3R5bGVzOiBbCiAgICAgIC8vIFNwZWNpYWwgaW1wb3J0IHJ1bGUgdG8gaW5jbHVkZSB0byBhbm90aGVyIGNzcyBmaWxlLgogICAgICBjc3MuaW1wb3J0KCdodHRwczovL2ZvbnRzLmdvb2dsZWFwaXMuY29tL2Nzcz9mYW1pbHk9Um9ib3RvJyksCiAgICAgIC8vIEVhY2ggc3R5bGUgcnVsZSB0YWtlcyBhIHZhbGlkIGNzcyBzZWxlY3RvciBhbmQgYSBzZXQgb2Ygc3R5bGVzLgogICAgICAvLyBTdHlsZXMgYXJlIGRlZmluZWQgdXNpbmcgdHlwZS1zYWZlIGNzcyBiaW5kaW5ncyBhbmQgY2FuIGJlIGZyZWVseSBjaGFpbmVkIGFuZCBuZXN0ZWQuCiAgICAgIGNzcygnaHRtbCwgYm9keScpLnN0eWxlcygKICAgICAgICB3aWR0aDogMTAwLnBlcmNlbnQsCiAgICAgICAgbWluSGVpZ2h0OiAxMDAudmgsCiAgICAgICAgcGFkZGluZzogUGFkZGluZy56ZXJvLAogICAgICAgIG1hcmdpbjogTWFyZ2luLnplcm8sCiAgICAgICAgZm9udEZhbWlseTogY29uc3QgRm9udEZhbWlseS5saXN0KFtGb250RmFtaWx5KCdSb2JvdG8nKSwgRm9udEZhbWlsaWVzLnNhbnNTZXJpZl0pLAogICAgICApLAogICAgICBjc3MoJ2gxJykuc3R5bGVzKAogICAgICAgIG1hcmdpbjogTWFyZ2luLnVuc2V0LAogICAgICAgIGZvbnRTaXplOiA0LnJlbSwKICAgICAgKSwKICAgIF0se3teaHlkcmF0aW9ufX0KICAgIGhlYWQ6IFsKICAgICAgLy8gTGlua3MgdG8gdGhlIGNvbXBpbGVkIGNsaWVudCBlbnRyeXBvaW50LgogICAgICBzY3JpcHQoZGVmZXI6IHRydWUsIHNyYzogJ21haW4uZGFydC5qcycsIFtdKSx7eyNmbHV0dGVyfX0KICAgICAgLy8gVGhlIGdlbmVyYXRlZCBmbHV0dGVyIG1hbmlmZXN0IGFuZCBib290c3RyYXAgc2NyaXB0LgogICAgICBsaW5rKHJlbDogJ21hbmlmZXN0JywgaHJlZjogJ21hbmlmZXN0Lmpzb24nKSwKICAgICAgc2NyaXB0KHNyYzogImZsdXR0ZXJfYm9vdHN0cmFwLmpzIiwgYXN5bmM6IHRydWUsIFtdKSx7ey9mbHV0dGVyfX0KICAgIF0se3svaHlkcmF0aW9ufX17eyNoeWRyYXRpb259fXt7I2ZsdXR0ZXJ9fQogICAgaGVhZDogWwogICAgICAvLyBUaGUgZ2VuZXJhdGVkIGZsdXR0ZXIgbWFuaWZlc3QgYW5kIGJvb3RzdHJhcCBzY3JpcHQuCiAgICAgIGxpbmsocmVsOiAnbWFuaWZlc3QnLCBocmVmOiAnbWFuaWZlc3QuanNvbicpLAogICAgICBzY3JpcHQoc3JjOiAiZmx1dHRlcl9ib290c3RyYXAuanMiLCBhc3luYzogdHJ1ZSwgW10pLAogICAgXSx7ey9mbHV0dGVyfX17ey9oeWRyYXRpb259fQogICAgYm9keTogQXBwKCksCiAgKSk7Cn0K",
+          "Ly8gVGhlIGVudHJ5cG9pbnQgZm9yIHRoZSAqKnNlcnZlcioqIGVudmlyb25tZW50LgovLwovLyBUaGUgW21haW5dIG1ldGhvZCB3aWxsIG9ubHkgYmUgZXhlY3V0ZWQgb24gdGhlIHNlcnZlciBkdXJpbmcgcHJlLXJlbmRlcmluZy4KLy8gVG8gcnVuIGNvZGUgb24gdGhlIGNsaWVudCwge3teaHlkcmF0aW9ufX1zZWUgW3dlYi9tYWluLmRhcnRde3svaHlkcmF0aW9ufX17eyNoeWRyYXRpb259fXVzZSB0aGUgQGNsaWVudCBhbm5vdGF0aW9ue3svaHlkcmF0aW9ufX0uCgovLyBTZXJ2ZXItc3BlY2lmaWMgamFzcHIgaW1wb3J0LgppbXBvcnQgJ3BhY2thZ2U6amFzcHIvc2VydmVyLmRhcnQnOwoKLy8gSW1wb3J0cyB0aGUgW0FwcF0gY29tcG9uZW50LgppbXBvcnQgJ2FwcC5kYXJ0JzsKCi8vIFRoaXMgZmlsZSBpcyBnZW5lcmF0ZWQgYXV0b21hdGljYWxseSBieSBKYXNwciwgZG8gbm90IHJlbW92ZSBvciBlZGl0LgppbXBvcnQgJ2phc3ByX29wdGlvbnMuZGFydCc7Cgp2b2lkIG1haW4oKSB7CiAgLy8gSW5pdGlhbGl6ZXMgdGhlIHNlcnZlciBlbnZpcm9ubWVudCB3aXRoIHRoZSBnZW5lcmF0ZWQgZGVmYXVsdCBvcHRpb25zLgogIEphc3ByLmluaXRpYWxpemVBcHAoCiAgICBvcHRpb25zOiBkZWZhdWx0SmFzcHJPcHRpb25zLAogICk7CgogIC8vIFN0YXJ0cyB0aGUgYXBwLgogIC8vCiAgLy8gW0RvY3VtZW50XSByZW5kZXJzIHRoZSByb290IGRvY3VtZW50IHN0cnVjdHVyZSAoPGh0bWw+LCA8aGVhZD4gYW5kIDxib2R5PikKICAvLyB3aXRoIHRoZSBwcm92aWRlZCBwYXJhbWV0ZXJzIGFuZCBjb21wb25lbnRzLgogIHJ1bkFwcChEb2N1bWVudCgKICAgIHRpdGxlOiAne3tuYW1lfX0nLAogICAgc3R5bGVzOiBbCiAgICAgIC8vIFNwZWNpYWwgaW1wb3J0IHJ1bGUgdG8gaW5jbHVkZSB0byBhbm90aGVyIGNzcyBmaWxlLgogICAgICBjc3MuaW1wb3J0KCdodHRwczovL2ZvbnRzLmdvb2dsZWFwaXMuY29tL2Nzcz9mYW1pbHk9Um9ib3RvJyksCiAgICAgIC8vIEVhY2ggc3R5bGUgcnVsZSB0YWtlcyBhIHZhbGlkIGNzcyBzZWxlY3RvciBhbmQgYSBzZXQgb2Ygc3R5bGVzLgogICAgICAvLyBTdHlsZXMgYXJlIGRlZmluZWQgdXNpbmcgdHlwZS1zYWZlIGNzcyBiaW5kaW5ncyBhbmQgY2FuIGJlIGZyZWVseSBjaGFpbmVkIGFuZCBuZXN0ZWQuCiAgICAgIGNzcygnaHRtbCwgYm9keScpLnN0eWxlcygKICAgICAgICB3aWR0aDogMTAwLnBlcmNlbnQsCiAgICAgICAgbWluSGVpZ2h0OiAxMDAudmgsCiAgICAgICAgcGFkZGluZzogUGFkZGluZy56ZXJvLAogICAgICAgIG1hcmdpbjogTWFyZ2luLnplcm8sCiAgICAgICAgZm9udEZhbWlseTogY29uc3QgRm9udEZhbWlseS5saXN0KFtGb250RmFtaWx5KCdSb2JvdG8nKSwgRm9udEZhbWlsaWVzLnNhbnNTZXJpZl0pLAogICAgICApLAogICAgICBjc3MoJ2gxJykuc3R5bGVzKAogICAgICAgIG1hcmdpbjogTWFyZ2luLnVuc2V0LAogICAgICAgIGZvbnRTaXplOiA0LnJlbSwKICAgICAgKSwKICAgIF0se3teaHlkcmF0aW9ufX0KICAgIGhlYWQ6IFsKICAgICAgLy8gTGlua3MgdG8gdGhlIGNvbXBpbGVkIGNsaWVudCBlbnRyeXBvaW50LgogICAgICBzY3JpcHQoZGVmZXI6IHRydWUsIHNyYzogJ21haW4uZGFydC5qcycpLHt7I2ZsdXR0ZXJ9fQogICAgICAvLyBUaGUgZ2VuZXJhdGVkIGZsdXR0ZXIgbWFuaWZlc3QgYW5kIGJvb3RzdHJhcCBzY3JpcHQuCiAgICAgIGxpbmsocmVsOiAnbWFuaWZlc3QnLCBocmVmOiAnbWFuaWZlc3QuanNvbicpLAogICAgICBzY3JpcHQoc3JjOiAiZmx1dHRlcl9ib290c3RyYXAuanMiLCBhc3luYzogdHJ1ZSkse3svZmx1dHRlcn19CiAgICBdLHt7L2h5ZHJhdGlvbn19e3sjaHlkcmF0aW9ufX17eyNmbHV0dGVyfX0KICAgIGhlYWQ6IFsKICAgICAgLy8gVGhlIGdlbmVyYXRlZCBmbHV0dGVyIG1hbmlmZXN0IGFuZCBib290c3RyYXAgc2NyaXB0LgogICAgICBsaW5rKHJlbDogJ21hbmlmZXN0JywgaHJlZjogJ21hbmlmZXN0Lmpzb24nKSwKICAgICAgc2NyaXB0KHNyYzogImZsdXR0ZXJfYm9vdHN0cmFwLmpzIiwgYXN5bmM6IHRydWUpLAogICAgXSx7ey9mbHV0dGVyfX17ey9oeWRyYXRpb259fQogICAgYm9keTogQXBwKCksCiAgKSk7Cn0K",
       "type": "text"
     },
     {
@@ -80,7 +80,7 @@ final scaffoldBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "lib/shelf#main.dart",
       "data":
-          "Ly8gVGhlIGVudHJ5cG9pbnQgZm9yIHRoZSAqKnNlcnZlcioqIGVudmlyb25tZW50LgovLwovLyBUaGUgW21haW5dIG1ldGhvZCB3aWxsIG9ubHkgYmUgZXhlY3V0ZWQgb24gdGhlIHNlcnZlciBkdXJpbmcgcHJlLXJlbmRlcmluZy4KLy8gVG8gcnVuIGNvZGUgb24gdGhlIGNsaWVudCwge3teaHlkcmF0aW9ufX1zZWUgW3dlYi9tYWluLmRhcnRde3svaHlkcmF0aW9ufX17eyNoeWRyYXRpb259fXVzZSB0aGUgQGNsaWVudCBhbm5vdGF0aW9ue3svaHlkcmF0aW9ufX0uCgppbXBvcnQgJ2RhcnQ6aW8nOwoKLy8gU2VydmVyLXNwZWNpZmljIGphc3ByIGltcG9ydC4KaW1wb3J0ICdwYWNrYWdlOmphc3ByL3NlcnZlci5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOnNoZWxmL3NoZWxmLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6c2hlbGYvc2hlbGZfaW8uZGFydCcgYXMgc2hlbGZfaW87CmltcG9ydCAncGFja2FnZTpzaGVsZl9yb3V0ZXIvc2hlbGZfcm91dGVyLmRhcnQnOwoKLy8gSW1wb3J0cyB0aGUgW0FwcF0gY29tcG9uZW50LgppbXBvcnQgJ2FwcC5kYXJ0JzsKCi8vIFRoaXMgZmlsZSBpcyBnZW5lcmF0ZWQgYXV0b21hdGljYWxseSBieSBKYXNwciwgZG8gbm90IHJlbW92ZSBvciBlZGl0LgppbXBvcnQgJ2phc3ByX29wdGlvbnMuZGFydCc7CgovLy8gSW5pdGlhbGl6ZXMgdGhlIGN1c3RvbSBzaGVsZiBzZXJ2ZXIuCi8vLwovLy8gVGhlIG1haW4oKSBmdW5jdGlvbiB3aWxsIGJlIGNhbGxlZCBhZ2FpbiBkdXJpbmcgZGV2ZWxvcG1lbnQgd2hlbiBob3QtcmVsb2FkaW5nLgovLy8gQ3VzdG9tIGJhY2tlbmQgaW1wbGVtZW50YXRpb25zIG11c3QgdGFrZSBjYXJlIG9mIHByb3Blcmx5IG1hbmFnaW5nIG9wZW4gaHR0cCBzZXJ2ZXJzCi8vLyBhbmQgb3RoZXIgcmVzb3VyY2VzIHRoYXQgbWlnaHQgYmUgcmUtY3JlYXRlZCB3aGVuIGhvdC1yZWxvYWRpbmcuCnZvaWQgbWFpbigpIGFzeW5jIHsKICBKYXNwci5pbml0aWFsaXplQXBwKAogICAgb3B0aW9uczogZGVmYXVsdEphc3ByT3B0aW9ucywKICApOwoKICB2YXIgcm91dGVyID0gUm91dGVyKCk7CgogIC8vIFJvdXRlIHlvdXIgYXBpIHJlcXVlc3RzIHRvIHlvdXIgb3duIGVuZHBvaW50LgogIHJvdXRlci5tb3VudCgnL2FwaScsIChyZXF1ZXN0KSB7CiAgICByZXR1cm4gUmVzcG9uc2Uub2soIkhlbGxvIEFwaSIpOwogIH0pOwoKICAvLyBVc2UgW3NlcnZlQXBwXSBpbnN0ZWFkIG9mIFtydW5BcHBdIHRvIGdldCBhIHNoZWxmIGhhbmRsZXIgeW91IGNhbiBtb3VudC4KICByb3V0ZXIubW91bnQoJy8nLCBzZXJ2ZUFwcCgocmVxdWVzdCwgcmVuZGVyKSB7CiAgICAvLyBPcHRpb25hbGx5IGRvIHNvbWV0aGluZyB3aXRoIGByZXF1ZXN0YAogICAgcHJpbnQoIlJlcXVlc3QgdXJpIGlzICR7cmVxdWVzdC5yZXF1ZXN0ZWRVcml9ICgke3JlcXVlc3QudXJsfSkiKTsKICAgIC8vIFJldHVybiBhIHNlcnZlci1yZW5kZXJlZCByZXNwb25zZSBieSBjYWxsaW5nIGByZW5kZXIoKWAgd2l0aCB5b3VyIHJvb3QgY29tcG9uZW50CiAgICByZXR1cm4gcmVuZGVyKERvY3VtZW50KAogICAgICB0aXRsZTogJ3t7bmFtZX19JywKICAgICAgc3R5bGVzOiBbCiAgICAgICAgLy8gU3BlY2lhbCBpbXBvcnQgcnVsZSB0byBpbmNsdWRlIHRvIGFub3RoZXIgY3NzIGZpbGUuCiAgICAgICAgY3NzLmltcG9ydCgnaHR0cHM6Ly9mb250cy5nb29nbGVhcGlzLmNvbS9jc3M/ZmFtaWx5PVJvYm90bycpLAogICAgICAgIC8vIEVhY2ggc3R5bGUgcnVsZSB0YWtlcyBhIHZhbGlkIGNzcyBzZWxlY3RvciBhbmQgYSBzZXQgb2Ygc3R5bGVzLgogICAgICAgIC8vIFN0eWxlcyBhcmUgZGVmaW5lZCB1c2luZyB0eXBlLXNhZmUgY3NzIGJpbmRpbmdzIGFuZCBjYW4gYmUgZnJlZWx5IGNoYWluZWQgYW5kIG5lc3RlZC4KICAgICAgICBjc3MoJ2h0bWwsIGJvZHknKS5zdHlsZXMoCiAgICAgICAgICB3aWR0aDogMTAwLnBlcmNlbnQsCiAgICAgICAgICBtaW5IZWlnaHQ6IDEwMC52aCwKICAgICAgICAgIHBhZGRpbmc6IFBhZGRpbmcuemVybywKICAgICAgICAgIG1hcmdpbjogTWFyZ2luLnplcm8sCiAgICAgICAgICBmb250RmFtaWx5OiBjb25zdCBGb250RmFtaWx5Lmxpc3QoW0ZvbnRGYW1pbHkoJ1JvYm90bycpLCBGb250RmFtaWxpZXMuc2Fuc1NlcmlmXSksCiAgICAgICAgKSwKICAgICAgICBjc3MoJ2gxJykuc3R5bGVzKAogICAgICAgICAgbWFyZ2luOiBNYXJnaW4udW5zZXQsCiAgICAgICAgICBmb250U2l6ZTogNC5yZW0sCiAgICAgICAgKSwKICAgICAgXSx7e15oeWRyYXRpb259fQogICAgICBoZWFkOiBbCiAgICAgICAgLy8gTGlua3MgdG8gdGhlIGNvbXBpbGVkIGNsaWVudCBlbnRyeXBvaW50LgogICAgICAgIHNjcmlwdChkZWZlcjogdHJ1ZSwgc3JjOiAnbWFpbi5kYXJ0LmpzJywgW10pLHt7I2ZsdXR0ZXJ9fQogICAgICAgIC8vIFRoZSBnZW5lcmF0ZWQgZmx1dHRlciBtYW5pZmVzdCBhbmQgYm9vdHN0cmFwIHNjcmlwdC4KICAgICAgICBsaW5rKHJlbDogJ21hbmlmZXN0JywgaHJlZjogJ21hbmlmZXN0Lmpzb24nKSwKICAgICAgICBzY3JpcHQoc3JjOiAiZmx1dHRlcl9ib290c3RyYXAuanMiLCBhc3luYzogdHJ1ZSwgW10pLHt7L2ZsdXR0ZXJ9fQogICAgICBdLHt7L2h5ZHJhdGlvbn19e3sjaHlkcmF0aW9ufX17eyNmbHV0dGVyfX0KICAgICAgaGVhZDogWwogICAgICAgIC8vIFRoZSBnZW5lcmF0ZWQgZmx1dHRlciBtYW5pZmVzdCBhbmQgYm9vdHN0cmFwIHNjcmlwdC4KICAgICAgICBsaW5rKHJlbDogJ21hbmlmZXN0JywgaHJlZjogJ21hbmlmZXN0Lmpzb24nKSwKICAgICAgICBzY3JpcHQoc3JjOiAiZmx1dHRlcl9ib290c3RyYXAuanMiLCBhc3luYzogdHJ1ZSwgW10pLAogICAgICBdLHt7L2ZsdXR0ZXJ9fXt7L2h5ZHJhdGlvbn19CiAgICAgIGJvZHk6IEFwcCgpLAogICAgKSk7CiAgfSkpOwoKICB2YXIgaGFuZGxlciA9IGNvbnN0IFBpcGVsaW5lKCkgLy8KICAgICAgLmFkZE1pZGRsZXdhcmUobG9nUmVxdWVzdHMoKSkKICAgICAgLmFkZEhhbmRsZXIocm91dGVyKTsKCiAgLy8gT2JqZWN0IHRvIHJlc29sdmUgYXN5bmMgbG9ja2luZyBvZiByZWxvYWRzLgogIHZhciByZWxvYWRMb2NrID0gYWN0aXZlUmVsb2FkTG9jayA9IE9iamVjdCgpOwoKICB2YXIgc2VydmVyID0gYXdhaXQgc2hlbGZfaW8uc2VydmUoaGFuZGxlciwgSW50ZXJuZXRBZGRyZXNzLmFueUlQdjQsIDgwODAsIHNoYXJlZDogdHJ1ZSk7CgogIC8vIElmIHRoZSByZWxvYWQgbG9jayBjaGFuZ2VkLCBhbm90aGVyIHJlbG9hZCBoYXBwZW5lZCBhbmQgd2Ugc2hvdWxkIGFib3J0LgogIGlmIChyZWxvYWRMb2NrICE9IGFjdGl2ZVJlbG9hZExvY2spIHsKICAgIHNlcnZlci5jbG9zZSgpOwogICAgcmV0dXJuOwogIH0KCiAgLy8gRWxzZSB3ZSBjYW4gc2FmZWx5IHVwZGF0ZSB0aGUgYWN0aXZlIHNlcnZlci4KICBhY3RpdmVTZXJ2ZXI/LmNsb3NlKCk7CiAgYWN0aXZlU2VydmVyID0gc2VydmVyOwoKICBwcmludCgnU2VydmluZyBhdCBodHRwOi8vJHtzZXJ2ZXIuYWRkcmVzcy5ob3N0fToke3NlcnZlci5wb3J0fScpOwp9CgovLy8gS2VlcHMgdHJhY2sgb2YgdGhlIGN1cnJlbnRseSBydW5uaW5nIGh0dHAgc2VydmVyLgpIdHRwU2VydmVyPyBhY3RpdmVTZXJ2ZXI7CgovLy8gS2VlcHMgdHJhY2sgb2YgdGhlIGxhc3QgY3JlYXRlZCByZWxvYWQgbG9jay4KLy8vIFRoaXMgaXMgbmVlZGVkIHRvIHRyYWNrIHJlbG9hZHMgd2hpY2ggbWlnaHQgaGFwcGVuIGluIHF1aWNrIHN1Y2Nlc3Npb24uCk9iamVjdD8gYWN0aXZlUmVsb2FkTG9jazsK",
+          "Ly8gVGhlIGVudHJ5cG9pbnQgZm9yIHRoZSAqKnNlcnZlcioqIGVudmlyb25tZW50LgovLwovLyBUaGUgW21haW5dIG1ldGhvZCB3aWxsIG9ubHkgYmUgZXhlY3V0ZWQgb24gdGhlIHNlcnZlciBkdXJpbmcgcHJlLXJlbmRlcmluZy4KLy8gVG8gcnVuIGNvZGUgb24gdGhlIGNsaWVudCwge3teaHlkcmF0aW9ufX1zZWUgW3dlYi9tYWluLmRhcnRde3svaHlkcmF0aW9ufX17eyNoeWRyYXRpb259fXVzZSB0aGUgQGNsaWVudCBhbm5vdGF0aW9ue3svaHlkcmF0aW9ufX0uCgppbXBvcnQgJ2RhcnQ6aW8nOwoKLy8gU2VydmVyLXNwZWNpZmljIGphc3ByIGltcG9ydC4KaW1wb3J0ICdwYWNrYWdlOmphc3ByL3NlcnZlci5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOnNoZWxmL3NoZWxmLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6c2hlbGYvc2hlbGZfaW8uZGFydCcgYXMgc2hlbGZfaW87CmltcG9ydCAncGFja2FnZTpzaGVsZl9yb3V0ZXIvc2hlbGZfcm91dGVyLmRhcnQnOwoKLy8gSW1wb3J0cyB0aGUgW0FwcF0gY29tcG9uZW50LgppbXBvcnQgJ2FwcC5kYXJ0JzsKCi8vIFRoaXMgZmlsZSBpcyBnZW5lcmF0ZWQgYXV0b21hdGljYWxseSBieSBKYXNwciwgZG8gbm90IHJlbW92ZSBvciBlZGl0LgppbXBvcnQgJ2phc3ByX29wdGlvbnMuZGFydCc7CgovLy8gSW5pdGlhbGl6ZXMgdGhlIGN1c3RvbSBzaGVsZiBzZXJ2ZXIuCi8vLwovLy8gVGhlIG1haW4oKSBmdW5jdGlvbiB3aWxsIGJlIGNhbGxlZCBhZ2FpbiBkdXJpbmcgZGV2ZWxvcG1lbnQgd2hlbiBob3QtcmVsb2FkaW5nLgovLy8gQ3VzdG9tIGJhY2tlbmQgaW1wbGVtZW50YXRpb25zIG11c3QgdGFrZSBjYXJlIG9mIHByb3Blcmx5IG1hbmFnaW5nIG9wZW4gaHR0cCBzZXJ2ZXJzCi8vLyBhbmQgb3RoZXIgcmVzb3VyY2VzIHRoYXQgbWlnaHQgYmUgcmUtY3JlYXRlZCB3aGVuIGhvdC1yZWxvYWRpbmcuCnZvaWQgbWFpbigpIGFzeW5jIHsKICBKYXNwci5pbml0aWFsaXplQXBwKAogICAgb3B0aW9uczogZGVmYXVsdEphc3ByT3B0aW9ucywKICApOwoKICB2YXIgcm91dGVyID0gUm91dGVyKCk7CgogIC8vIFJvdXRlIHlvdXIgYXBpIHJlcXVlc3RzIHRvIHlvdXIgb3duIGVuZHBvaW50LgogIHJvdXRlci5tb3VudCgnL2FwaScsIChyZXF1ZXN0KSB7CiAgICByZXR1cm4gUmVzcG9uc2Uub2soIkhlbGxvIEFwaSIpOwogIH0pOwoKICAvLyBVc2UgW3NlcnZlQXBwXSBpbnN0ZWFkIG9mIFtydW5BcHBdIHRvIGdldCBhIHNoZWxmIGhhbmRsZXIgeW91IGNhbiBtb3VudC4KICByb3V0ZXIubW91bnQoJy8nLCBzZXJ2ZUFwcCgocmVxdWVzdCwgcmVuZGVyKSB7CiAgICAvLyBPcHRpb25hbGx5IGRvIHNvbWV0aGluZyB3aXRoIGByZXF1ZXN0YAogICAgcHJpbnQoIlJlcXVlc3QgdXJpIGlzICR7cmVxdWVzdC5yZXF1ZXN0ZWRVcml9ICgke3JlcXVlc3QudXJsfSkiKTsKICAgIC8vIFJldHVybiBhIHNlcnZlci1yZW5kZXJlZCByZXNwb25zZSBieSBjYWxsaW5nIGByZW5kZXIoKWAgd2l0aCB5b3VyIHJvb3QgY29tcG9uZW50CiAgICByZXR1cm4gcmVuZGVyKERvY3VtZW50KAogICAgICB0aXRsZTogJ3t7bmFtZX19JywKICAgICAgc3R5bGVzOiBbCiAgICAgICAgLy8gU3BlY2lhbCBpbXBvcnQgcnVsZSB0byBpbmNsdWRlIHRvIGFub3RoZXIgY3NzIGZpbGUuCiAgICAgICAgY3NzLmltcG9ydCgnaHR0cHM6Ly9mb250cy5nb29nbGVhcGlzLmNvbS9jc3M/ZmFtaWx5PVJvYm90bycpLAogICAgICAgIC8vIEVhY2ggc3R5bGUgcnVsZSB0YWtlcyBhIHZhbGlkIGNzcyBzZWxlY3RvciBhbmQgYSBzZXQgb2Ygc3R5bGVzLgogICAgICAgIC8vIFN0eWxlcyBhcmUgZGVmaW5lZCB1c2luZyB0eXBlLXNhZmUgY3NzIGJpbmRpbmdzIGFuZCBjYW4gYmUgZnJlZWx5IGNoYWluZWQgYW5kIG5lc3RlZC4KICAgICAgICBjc3MoJ2h0bWwsIGJvZHknKS5zdHlsZXMoCiAgICAgICAgICB3aWR0aDogMTAwLnBlcmNlbnQsCiAgICAgICAgICBtaW5IZWlnaHQ6IDEwMC52aCwKICAgICAgICAgIHBhZGRpbmc6IFBhZGRpbmcuemVybywKICAgICAgICAgIG1hcmdpbjogTWFyZ2luLnplcm8sCiAgICAgICAgICBmb250RmFtaWx5OiBjb25zdCBGb250RmFtaWx5Lmxpc3QoW0ZvbnRGYW1pbHkoJ1JvYm90bycpLCBGb250RmFtaWxpZXMuc2Fuc1NlcmlmXSksCiAgICAgICAgKSwKICAgICAgICBjc3MoJ2gxJykuc3R5bGVzKAogICAgICAgICAgbWFyZ2luOiBNYXJnaW4udW5zZXQsCiAgICAgICAgICBmb250U2l6ZTogNC5yZW0sCiAgICAgICAgKSwKICAgICAgXSx7e15oeWRyYXRpb259fQogICAgICBoZWFkOiBbCiAgICAgICAgLy8gTGlua3MgdG8gdGhlIGNvbXBpbGVkIGNsaWVudCBlbnRyeXBvaW50LgogICAgICAgIHNjcmlwdChkZWZlcjogdHJ1ZSwgc3JjOiAnbWFpbi5kYXJ0LmpzJykse3sjZmx1dHRlcn19CiAgICAgICAgLy8gVGhlIGdlbmVyYXRlZCBmbHV0dGVyIG1hbmlmZXN0IGFuZCBib290c3RyYXAgc2NyaXB0LgogICAgICAgIGxpbmsocmVsOiAnbWFuaWZlc3QnLCBocmVmOiAnbWFuaWZlc3QuanNvbicpLAogICAgICAgIHNjcmlwdChzcmM6ICJmbHV0dGVyX2Jvb3RzdHJhcC5qcyIsIGFzeW5jOiB0cnVlKSx7ey9mbHV0dGVyfX0KICAgICAgXSx7ey9oeWRyYXRpb259fXt7I2h5ZHJhdGlvbn19e3sjZmx1dHRlcn19CiAgICAgIGhlYWQ6IFsKICAgICAgICAvLyBUaGUgZ2VuZXJhdGVkIGZsdXR0ZXIgbWFuaWZlc3QgYW5kIGJvb3RzdHJhcCBzY3JpcHQuCiAgICAgICAgbGluayhyZWw6ICdtYW5pZmVzdCcsIGhyZWY6ICdtYW5pZmVzdC5qc29uJyksCiAgICAgICAgc2NyaXB0KHNyYzogImZsdXR0ZXJfYm9vdHN0cmFwLmpzIiwgYXN5bmM6IHRydWUpLAogICAgICBdLHt7L2ZsdXR0ZXJ9fXt7L2h5ZHJhdGlvbn19CiAgICAgIGJvZHk6IEFwcCgpLAogICAgKSk7CiAgfSkpOwoKICB2YXIgaGFuZGxlciA9IGNvbnN0IFBpcGVsaW5lKCkgLy8KICAgICAgLmFkZE1pZGRsZXdhcmUobG9nUmVxdWVzdHMoKSkKICAgICAgLmFkZEhhbmRsZXIocm91dGVyKTsKCiAgLy8gT2JqZWN0IHRvIHJlc29sdmUgYXN5bmMgbG9ja2luZyBvZiByZWxvYWRzLgogIHZhciByZWxvYWRMb2NrID0gYWN0aXZlUmVsb2FkTG9jayA9IE9iamVjdCgpOwoKICB2YXIgc2VydmVyID0gYXdhaXQgc2hlbGZfaW8uc2VydmUoaGFuZGxlciwgSW50ZXJuZXRBZGRyZXNzLmFueUlQdjQsIDgwODAsIHNoYXJlZDogdHJ1ZSk7CgogIC8vIElmIHRoZSByZWxvYWQgbG9jayBjaGFuZ2VkLCBhbm90aGVyIHJlbG9hZCBoYXBwZW5lZCBhbmQgd2Ugc2hvdWxkIGFib3J0LgogIGlmIChyZWxvYWRMb2NrICE9IGFjdGl2ZVJlbG9hZExvY2spIHsKICAgIHNlcnZlci5jbG9zZSgpOwogICAgcmV0dXJuOwogIH0KCiAgLy8gRWxzZSB3ZSBjYW4gc2FmZWx5IHVwZGF0ZSB0aGUgYWN0aXZlIHNlcnZlci4KICBhY3RpdmVTZXJ2ZXI/LmNsb3NlKCk7CiAgYWN0aXZlU2VydmVyID0gc2VydmVyOwoKICBwcmludCgnU2VydmluZyBhdCBodHRwOi8vJHtzZXJ2ZXIuYWRkcmVzcy5ob3N0fToke3NlcnZlci5wb3J0fScpOwp9CgovLy8gS2VlcHMgdHJhY2sgb2YgdGhlIGN1cnJlbnRseSBydW5uaW5nIGh0dHAgc2VydmVyLgpIdHRwU2VydmVyPyBhY3RpdmVTZXJ2ZXI7CgovLy8gS2VlcHMgdHJhY2sgb2YgdGhlIGxhc3QgY3JlYXRlZCByZWxvYWQgbG9jay4KLy8vIFRoaXMgaXMgbmVlZGVkIHRvIHRyYWNrIHJlbG9hZHMgd2hpY2ggbWlnaHQgaGFwcGVuIGluIHF1aWNrIHN1Y2Nlc3Npb24uCk9iamVjdD8gYWN0aXZlUmVsb2FkTG9jazsK",
       "type": "text"
     },
     {
