+++ packages/jaspr_cli/lib/src/bundles/templates/docs/docs_bundle.dart
@@ -56,7 +56,7 @@ final docsBundle = MasonBundle.fromJson(<String, dynamic>{
     {
       "path": "lib/main.dart",
       "data":
-          "Ly8gVGhlIGVudHJ5cG9pbnQgZm9yIHRoZSAqKnNlcnZlcioqIGVudmlyb25tZW50LgovLwovLyBUaGUgW21haW5dIG1ldGhvZCB3aWxsIG9ubHkgYmUgZXhlY3V0ZWQgb24gdGhlIHNlcnZlciBkdXJpbmcgcHJlLXJlbmRlcmluZy4KLy8gVG8gcnVuIGNvZGUgb24gdGhlIGNsaWVudCwgdXNlIHRoZSBAY2xpZW50IGFubm90YXRpb24uCgovLyBTZXJ2ZXItc3BlY2lmaWMgamFzcHIgaW1wb3J0LgppbXBvcnQgJ3BhY2thZ2U6amFzcHIvc2VydmVyLmRhcnQnOwoKaW1wb3J0ICdwYWNrYWdlOmphc3ByX2NvbnRlbnQvamFzcHJfY29udGVudC5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOmphc3ByX2NvbnRlbnQvY29tcG9uZW50cy9jb2RlX2Jsb2NrLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6amFzcHJfY29udGVudC9jb21wb25lbnRzL3RoZW1lX3RvZ2dsZS5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOmphc3ByX2NvbnRlbnQvY29tcG9uZW50cy9naXRodWJfYnV0dG9uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6amFzcHJfY29udGVudC9jb21wb25lbnRzL2NhbGxvdXQuZGFydCc7CmltcG9ydCAncGFja2FnZTpqYXNwcl9jb250ZW50L2NvbXBvbmVudHMvaW1hZ2UuZGFydCc7CmltcG9ydCAncGFja2FnZTpqYXNwcl9jb250ZW50L3RoZW1lLmRhcnQnOwoKaW1wb3J0ICdjb21wb25lbnRzL2NsaWNrZXIuZGFydCc7CgovLyBUaGlzIGZpbGUgaXMgZ2VuZXJhdGVkIGF1dG9tYXRpY2FsbHkgYnkgSmFzcHIsIGRvIG5vdCByZW1vdmUgb3IgZWRpdC4KaW1wb3J0ICdqYXNwcl9vcHRpb25zLmRhcnQnOwoKdm9pZCBtYWluKCkgewogIC8vIEluaXRpYWxpemVzIHRoZSBzZXJ2ZXIgZW52aXJvbm1lbnQgd2l0aCB0aGUgZ2VuZXJhdGVkIGRlZmF1bHQgb3B0aW9ucy4KICBKYXNwci5pbml0aWFsaXplQXBwKAogICAgb3B0aW9uczogZGVmYXVsdEphc3ByT3B0aW9ucywKICApOwoKICAvLyBTdGFydHMgdGhlIGFwcC4KICAvLwogIC8vIFtDb250ZW50QXBwXSBzcGlucyB1cCB0aGUgY29udGVudCByZW5kZXJpbmcgcGlwZWxpbmUgZnJvbSBqYXNwcl9jb250ZW50IHRvIHJlbmRlcgogIC8vIHlvdXIgbWFya2Rvd24gZmlsZXMgaW4gdGhlIGNvbnRlbnQvIGRpcmVjdG9yeSB0byBhIGJlYXV0aWZ1bCBkb2N1bWVudGF0aW9uIHNpdGUuCiAgcnVuQXBwKENvbnRlbnRBcHAoCiAgICAvLyBFbmFibGVzIG11c3RhY2hlIHRlbXBsYXRpbmcgaW5zaWRlIHRoZSBtYXJrZG93biBmaWxlcy4KICAgIHRlbXBsYXRlRW5naW5lOiBNdXN0YWNoZVRlbXBsYXRlRW5naW5lKCksCiAgICBwYXJzZXJzOiBbCiAgICAgIE1hcmtkb3duUGFyc2VyKCksCiAgICBdLAogICAgZXh0ZW5zaW9uczogWwogICAgICAvLyBBZGRzIGhlYWRpbmcgYW5jaG9ycyB0byBlYWNoIGhlYWRpbmcuCiAgICAgIEhlYWRpbmdBbmNob3JzRXh0ZW5zaW9uKCksCiAgICAgIC8vIEdlbmVyYXRlcyBhIHRhYmxlIG9mIGNvbnRlbnRzIGZvciBlYWNoIHBhZ2UuCiAgICAgIFRhYmxlT2ZDb250ZW50c0V4dGVuc2lvbigpLAogICAgXSwKICAgIGNvbXBvbmVudHM6IFsKICAgICAgLy8gVGhlIDxJbmZvPiBibG9jayBhbmQgb3RoZXIgY2FsbG91dHMuCiAgICAgIENhbGxvdXQoKSwKICAgICAgLy8gQWRkcyBzeW50YXggaGlnaGxpZ2h0aW5nIHRvIGNvZGUgYmxvY2tzLgogICAgICBDb2RlQmxvY2soKSwKICAgICAgLy8gQWRkcyBhIGN1c3RvbSBKYXNwciBjb21wb25lbnQgdG8gYmUgdXNlZCBhcyA8Q2xpY2tlci8+IGluIG1hcmtkb3duLgogICAgICBDdXN0b21Db21wb25lbnQoCiAgICAgICAgcGF0dGVybjogJ0NsaWNrZXInLAogICAgICAgIGJ1aWxkZXI6IChfLCBfXywgX19fKSA9PiBDbGlja2VyKCksCiAgICAgICksCiAgICAgIC8vIEFkZHMgem9vbWluZyBhbmQgY2FwdGlvbiBzdXBwb3J0IHRvIGltYWdlcy4KICAgICAgSW1hZ2Uoem9vbTogdHJ1ZSksCiAgICBdLAogICAgbGF5b3V0czogWwogICAgICAvLyBPdXQtb2YtdGhlLWJveCBsYXlvdXQgZm9yIGRvY3VtZW50YXRpb24gc2l0ZXMuCiAgICAgIERvY3NMYXlvdXQoCiAgICAgICAgaGVhZGVyOiBIZWFkZXIoCiAgICAgICAgICB0aXRsZTogJ015IERvY3MnLAogICAgICAgICAgbG9nbzogJy9pbWFnZXMvbG9nby5zdmcnLAogICAgICAgICAgaXRlbXM6IFsKICAgICAgICAgICAgLy8gRW5hYmxlcyBzd2l0Y2hpbmcgYmV0d2VlbiBsaWdodCBhbmQgZGFyayBtb2RlLgogICAgICAgICAgICBUaGVtZVRvZ2dsZSgpLAogICAgICAgICAgICAvLyBTaG93cyBnaXRodWIgc3RhdHMuCiAgICAgICAgICAgIEdpdGh1YkJ1dHRvbihyZXBvOiAnc2NodWx0ZWsvamFzcHInKSwKICAgICAgICAgIF0sCiAgICAgICAgKSwKICAgICAgICBzaWRlYmFyOiBTaWRlYmFyKGdyb3VwczogWwogICAgICAgICAgLy8gQWRkcyBuYXZpZ2F0aW9uIGxpbmtzIHRvIHRoZSBzaWRlYmFyLgogICAgICAgICAgU2lkZWJhckdyb3VwKAogICAgICAgICAgICBsaW5rczogWwogICAgICAgICAgICAgIFNpZGViYXJMaW5rKHRleHQ6ICJPdmVydmlldyIsIGhyZWY6ICcvJyksCiAgICAgICAgICAgIF0sCiAgICAgICAgICApLAogICAgICAgICAgU2lkZWJhckdyb3VwKHRpdGxlOiAnQ29udGVudCcsIGxpbmtzOiBbCiAgICAgICAgICAgIFNpZGViYXJMaW5rKHRleHQ6ICJBYm91dCIsIGhyZWY6ICcvYWJvdXQnKSwKICAgICAgICAgIF0pLAogICAgICAgIF0pLAogICAgICApLAogICAgXSwKICAgIHRoZW1lOiBDb250ZW50VGhlbWUoCiAgICAgIC8vIEN1c3RvbWl6ZXMgdGhlIGRlZmF1bHQgdGhlbWUgY29sb3JzLgogICAgICBwcmltYXJ5OiBUaGVtZUNvbG9yKFRoZW1lQ29sb3JzLmJsdWUuJDUwMCwgZGFyazogVGhlbWVDb2xvcnMuYmx1ZS4kMzAwKSwKICAgICAgYmFja2dyb3VuZDogVGhlbWVDb2xvcihUaGVtZUNvbG9ycy5zbGF0ZS4kNTAsIGRhcms6IFRoZW1lQ29sb3JzLnppbmMuJDk1MCksCiAgICAgIGNvbG9yczogWwogICAgICAgIENvbnRlbnRDb2xvcnMucXVvdGVCb3JkZXJzLmFwcGx5KFRoZW1lQ29sb3JzLmJsdWUuJDQwMCksCiAgICAgIF0sCiAgICApLAogICkpOwp9Cg==",
+          "Ly8gVGhlIGVudHJ5cG9pbnQgZm9yIHRoZSAqKnNlcnZlcioqIGVudmlyb25tZW50LgovLwovLyBUaGUgW21haW5dIG1ldGhvZCB3aWxsIG9ubHkgYmUgZXhlY3V0ZWQgb24gdGhlIHNlcnZlciBkdXJpbmcgcHJlLXJlbmRlcmluZy4KLy8gVG8gcnVuIGNvZGUgb24gdGhlIGNsaWVudCwgdXNlIHRoZSBAY2xpZW50IGFubm90YXRpb24uCgovLyBTZXJ2ZXItc3BlY2lmaWMgamFzcHIgaW1wb3J0LgppbXBvcnQgJ3BhY2thZ2U6amFzcHIvc2VydmVyLmRhcnQnOwoKaW1wb3J0ICdwYWNrYWdlOmphc3ByX2NvbnRlbnQvamFzcHJfY29udGVudC5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOmphc3ByX2NvbnRlbnQvY29tcG9uZW50cy9jb2RlX2Jsb2NrLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6amFzcHJfY29udGVudC9jb21wb25lbnRzL3RoZW1lX3RvZ2dsZS5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOmphc3ByX2NvbnRlbnQvY29tcG9uZW50cy9naXRodWJfYnV0dG9uLmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6amFzcHJfY29udGVudC9jb21wb25lbnRzL2NhbGxvdXQuZGFydCc7CmltcG9ydCAncGFja2FnZTpqYXNwcl9jb250ZW50L2NvbXBvbmVudHMvaW1hZ2UuZGFydCc7CmltcG9ydCAncGFja2FnZTpqYXNwcl9jb250ZW50L3RoZW1lLmRhcnQnOwoKaW1wb3J0ICdjb21wb25lbnRzL2NsaWNrZXIuZGFydCc7CgovLyBUaGlzIGZpbGUgaXMgZ2VuZXJhdGVkIGF1dG9tYXRpY2FsbHkgYnkgSmFzcHIsIGRvIG5vdCByZW1vdmUgb3IgZWRpdC4KaW1wb3J0ICdqYXNwcl9vcHRpb25zLmRhcnQnOwoKdm9pZCBtYWluKCkgewogIC8vIEluaXRpYWxpemVzIHRoZSBzZXJ2ZXIgZW52aXJvbm1lbnQgd2l0aCB0aGUgZ2VuZXJhdGVkIGRlZmF1bHQgb3B0aW9ucy4KICBKYXNwci5pbml0aWFsaXplQXBwKAogICAgb3B0aW9uczogZGVmYXVsdEphc3ByT3B0aW9ucywKICApOwoKICAvLyBTdGFydHMgdGhlIGFwcC4KICAvLwogIC8vIFtDb250ZW50QXBwXSBzcGlucyB1cCB0aGUgY29udGVudCByZW5kZXJpbmcgcGlwZWxpbmUgZnJvbSBqYXNwcl9jb250ZW50IHRvIHJlbmRlcgogIC8vIHlvdXIgbWFya2Rvd24gZmlsZXMgaW4gdGhlIGNvbnRlbnQvIGRpcmVjdG9yeSB0byBhIGJlYXV0aWZ1bCBkb2N1bWVudGF0aW9uIHNpdGUuCiAgcnVuQXBwKENvbnRlbnRBcHAoCiAgICAvLyBFbmFibGVzIG11c3RhY2hlIHRlbXBsYXRpbmcgaW5zaWRlIHRoZSBtYXJrZG93biBmaWxlcy4KICAgIHRlbXBsYXRlRW5naW5lOiBNdXN0YWNoZVRlbXBsYXRlRW5naW5lKCksCiAgICBwYXJzZXJzOiBbCiAgICAgIE1hcmtkb3duUGFyc2VyKCksCiAgICBdLAogICAgZXh0ZW5zaW9uczogWwogICAgICAvLyBBZGRzIGhlYWRpbmcgYW5jaG9ycyB0byBlYWNoIGhlYWRpbmcuCiAgICAgIEhlYWRpbmdBbmNob3JzRXh0ZW5zaW9uKCksCiAgICAgIC8vIEdlbmVyYXRlcyBhIHRhYmxlIG9mIGNvbnRlbnRzIGZvciBlYWNoIHBhZ2UuCiAgICAgIFRhYmxlT2ZDb250ZW50c0V4dGVuc2lvbigpLAogICAgXSwKICAgIGNvbXBvbmVudHM6IFsKICAgICAgLy8gVGhlIDxJbmZvPiBibG9jayBhbmQgb3RoZXIgY2FsbG91dHMuCiAgICAgIENhbGxvdXQoKSwKICAgICAgLy8gQWRkcyBzeW50YXggaGlnaGxpZ2h0aW5nIHRvIGNvZGUgYmxvY2tzLgogICAgICBDb2RlQmxvY2soKSwKICAgICAgLy8gQWRkcyBhIGN1c3RvbSBKYXNwciBjb21wb25lbnQgdG8gYmUgdXNlZCBhcyA8Q2xpY2tlci8+IGluIG1hcmtkb3duLgogICAgICBDdXN0b21Db21wb25lbnQoCiAgICAgICAgcGF0dGVybjogJ0NsaWNrZXInLAogICAgICAgIGJ1aWxkZXI6IChfLCBfXywgX19fKSA9PiBDbGlja2VyKCksCiAgICAgICksCiAgICAgIC8vIEFkZHMgem9vbWluZyBhbmQgY2FwdGlvbiBzdXBwb3J0IHRvIGltYWdlcy4KICAgICAgSW1hZ2Uoem9vbTogdHJ1ZSksCiAgICBdLAogICAgbGF5b3V0czogWwogICAgICAvLyBPdXQtb2YtdGhlLWJveCBsYXlvdXQgZm9yIGRvY3VtZW50YXRpb24gc2l0ZXMuCiAgICAgIERvY3NMYXlvdXQoCiAgICAgICAgaGVhZGVyOiBIZWFkZXIoCiAgICAgICAgICB0aXRsZTogJ015IERvY3MnLAogICAgICAgICAgbG9nbzogJy9pbWFnZXMvbG9nby5zdmcnLAogICAgICAgICAgaXRlbXM6IFsKICAgICAgICAgICAgLy8gRW5hYmxlcyBzd2l0Y2hpbmcgYmV0d2VlbiBsaWdodCBhbmQgZGFyayBtb2RlLgogICAgICAgICAgICBUaGVtZVRvZ2dsZSgpLAogICAgICAgICAgICAvLyBTaG93cyBnaXRodWIgc3RhdHMuCiAgICAgICAgICAgIEdpdEh1YkJ1dHRvbihyZXBvOiAnc2NodWx0ZWsvamFzcHInKSwKICAgICAgICAgIF0sCiAgICAgICAgKSwKICAgICAgICBzaWRlYmFyOiBTaWRlYmFyKGdyb3VwczogWwogICAgICAgICAgLy8gQWRkcyBuYXZpZ2F0aW9uIGxpbmtzIHRvIHRoZSBzaWRlYmFyLgogICAgICAgICAgU2lkZWJhckdyb3VwKAogICAgICAgICAgICBsaW5rczogWwogICAgICAgICAgICAgIFNpZGViYXJMaW5rKHRleHQ6ICJPdmVydmlldyIsIGhyZWY6ICcvJyksCiAgICAgICAgICAgIF0sCiAgICAgICAgICApLAogICAgICAgICAgU2lkZWJhckdyb3VwKHRpdGxlOiAnQ29udGVudCcsIGxpbmtzOiBbCiAgICAgICAgICAgIFNpZGViYXJMaW5rKHRleHQ6ICJBYm91dCIsIGhyZWY6ICcvYWJvdXQnKSwKICAgICAgICAgIF0pLAogICAgICAgIF0pLAogICAgICApLAogICAgXSwKICAgIHRoZW1lOiBDb250ZW50VGhlbWUoCiAgICAgIC8vIEN1c3RvbWl6ZXMgdGhlIGRlZmF1bHQgdGhlbWUgY29sb3JzLgogICAgICBwcmltYXJ5OiBUaGVtZUNvbG9yKFRoZW1lQ29sb3JzLmJsdWUuJDUwMCwgZGFyazogVGhlbWVDb2xvcnMuYmx1ZS4kMzAwKSwKICAgICAgYmFja2dyb3VuZDogVGhlbWVDb2xvcihUaGVtZUNvbG9ycy5zbGF0ZS4kNTAsIGRhcms6IFRoZW1lQ29sb3JzLnppbmMuJDk1MCksCiAgICAgIGNvbG9yczogWwogICAgICAgIENvbnRlbnRDb2xvcnMucXVvdGVCb3JkZXJzLmFwcGx5KFRoZW1lQ29sb3JzLmJsdWUuJDQwMCksCiAgICAgIF0sCiAgICApLAogICkpOwp9Cg==",
       "type": "text"
     },
     {
