+++ packages/jaspr_cli/tool/scaffold/__brick__/lib/base#main.dart
@@ -43,15 +43,15 @@ void main() {
     ],{{^hydration}}
     head: [
       // Links to the compiled client entrypoint.
-      script(defer: true, src: 'main.dart.js', []),{{#flutter}}
+      script(defer: true, src: 'main.dart.js'),{{#flutter}}
       // The generated flutter manifest and bootstrap script.
       link(rel: 'manifest', href: 'manifest.json'),
-      script(src: "flutter_bootstrap.js", async: true, []),{{/flutter}}
+      script(src: "flutter_bootstrap.js", async: true),{{/flutter}}
     ],{{/hydration}}{{#hydration}}{{#flutter}}
     head: [
       // The generated flutter manifest and bootstrap script.
       link(rel: 'manifest', href: 'manifest.json'),
-      script(src: "flutter_bootstrap.js", async: true, []),
+      script(src: "flutter_bootstrap.js", async: true),
     ],{{/flutter}}{{/hydration}}
     body: App(),
   ));
